+++ FrontEnd/angular12/src/app/app-routing.module.ts
@@ -18,6 +18,11 @@ import {LoginComponent} from "./screens/login/login.component";
 
 const routes: Routes = [
   {
+    path:'',
+    redirectTo: '/login',
+    pathMatch: 'full'
+  },
+  {
     path: 'login',
     component: LoginComponent
   },
