+++ FrontEnd/angular12/src/app/header/header.component.html
@@ -1,7 +1,7 @@
 <nav class="wrapperHeader">
 
   <div >
-  <a href="#">
+  <a href="registro">
   <img width="50" height="50" src="https://images.unsplash.com/photo-1637354563042-9a1d26500ff5?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=MnwxfDB8MXxyYW5kb218MHx8fHx8fHx8MTYzNzk4NjMyMw&ixlib=rb-1.2.1&q=80&w=1080">
   </a>
   <ul>
