+++ FrontEnd/angular12/src/app/screens/register/register.component.css
@@ -70,8 +70,6 @@
 .texts {
   display: grid;
   grid-template-columns: repeat(2, 1fr);
-
-
   width: 100%;
   height: 100%;
 }
@@ -84,26 +82,18 @@
 }
 
 .error {
-
   grid-column: 2;
   grid-row: 1;
   color: #f61111;
-
-
 }
 
 .buttonSubmit {
-
   grid-column: 1;
-
-
   background-color: rgba(255, 99, 99, 0.89);
   vertical-align: center;
   horiz-align: center;
-
   border: none;
   color: #ffffff;
-
   text-align: center;
   text-decoration: none;
   display: inline-block;
@@ -115,7 +105,6 @@
 .subText {
   grid-column: 1 /3;
   padding-left: 20px;
-
 }
