+++ FrontEnd/angular12/src/app/screens/register/register.component.html
@@ -10,16 +10,11 @@
       <h2 class="subText">Ingresar datos</h2>
 
       <div class="container">
-
         <div class="texts">
           <div class="name">Nombre del CCFS</div>
           <strong class="error" *ngIf="isBoxValid('ccfs')">Este campo es requerido</strong>
         </div>
-
-
         <input class="box" type="text" formControlName="ccfs"/>
-
-
       </div>
 
       <div class="container">
