+++ FrontEnd/angular12/src/app/screens/register/register.component.ts
@@ -64,9 +64,6 @@ export class RegisterComponent implements OnInit {
 
 
     });
-
-
-
   }
 
   places: string[] = [
